+++ src/main/java/com/uwetrottmann/trakt5/entities/LastActivityEpisodes.java
@@ -1,5 +1,4 @@
 /*
- * Copyright 2024 ZelKami
  * Copyright 2026 Uwe Trottmann <uwe@uwetrottmann.com>
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -19,10 +18,10 @@ package com.uwetrottmann.trakt5.entities;
 
 import org.threeten.bp.OffsetDateTime;
 
-public class LastActivityComments {
+public class LastActivityEpisodes extends LastActivity {
 
-    public OffsetDateTime liked_at;
-    public OffsetDateTime reacted_at;
-    public OffsetDateTime blocked_at;
+    public OffsetDateTime watched_at;
+    public OffsetDateTime collected_at;
+    public OffsetDateTime paused_at;
 
 }